# Only build the RTTI hashmap once per thread